Upgrade type-detect to v4 (#22881)
